Avoid bailout when PROFILE_SOURCED has been set.
This was originally done to avoid repeating calls to pathprepend and
growing the PATH with redundant entries; but it turns out to be a bad
idea when using tmux.  Tmux will spawn a new shell with none of the
functions found in .bashrc, but with the exported variable
PROFILE_SOURCED still set.